backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
geopandas
shapely
openpyxl
osmnx
